refactor: checkWorkspaceSettingsUpdatable command and related utilities
