+++ src/command/command/index.mts
@@ -1,3 +1,4 @@
+import checkWorkspaceSettingsUpdatable from "./checkWorkspaceSettingsUpdatable/index.mjs";
 import collectAdmonitions from "./collectAdmonitions/index.mjs";
 import createNote from "./createNote/index.mjs";
 import createWorkspaceFleximarkCss from "./createWorkspaceFleximarkCss/index.mjs";
@@ -6,6 +7,7 @@ import resetGlobalFleximarkCss from "./resetGlobalFleximarkCss/index.mjs";
 import updateWorkspaceSettings from "./updateWorkspaceSettings/index.mjs";
 
 export {
+  checkWorkspaceSettingsUpdatable,
   collectAdmonitions,
   createNote,
   createWorkspaceFleximarkCss,
